Scinde des gros fichiers du module member
